feat: add ColGroup to adjust column width
