+++ src/utils/buttonOptions.js
@@ -1,3 +1,5 @@
+// value는 setOrder()에 넣어줄 값
+// label은 dropdown 버튼에 보여질 부분
 const typeOneOptions = [
   { value: "highestInvestment", label: "누적 투자금액 높은순" },
   { value: "lowestInvestment", label: "누적 투자금액 낮은순" },
@@ -7,6 +9,9 @@ const typeOneOptions = [
   { value: "lowestEmployees", label: "고용 인원 적은순" },
 ];
 
+//orderBy하나로 상태 관리 하기때문에 fetch에 필요한 요청쿼리 2개를 이어줌
+//호출 Api에서 나눠줄 예정
+// ?orderBy=selectedCount&scending=asc
 const typeTwoOptions = [
   { value: "selectedCount_desc", label: "나의 기업 선택 횟수 높은순" },
   { value: "selectedCount_asc", label: "나의 기업 선택 횟수 낮은순" },
@@ -14,6 +19,7 @@ const typeTwoOptions = [
   { value: "actualInvestment_asc", label: "실제 누적 투자 금액 낮은순" },
 ];
 
+//?sortBy=virtualInvestment?order=desc
 const typeThreeOptions = [
   {
     value: "virtualInvestment_desc",
@@ -24,6 +30,9 @@ const typeThreeOptions = [
   { value: "actualInvestment_asc", label: "실제 누적 투자 금액 낮은순" },
 ];
 
+// 더 필요한 드롭다운 있다면 type{넘버}Options로 정의 후 optionsByType에 추가.
+// defaultSort = 기본 정렬 (페이지 맨 처음에 fetch될 기본 정렬 이름)
+//typeOne 은 small / typeTwo, typeThree는 large 여야함
 export const optionsByType = {
   typeOne: {
     options: typeOneOptions,
